==> 01_Llibreries.R <==
## Aniol Vilamala Vidal

################################################################################
# ----- LLIBRERIES EMPRADES EN EL PROJECTE -----
################################################################################
 
library(tidyverse) # Carrega paquets readr, dplyr, tidyr, tibble, forcats, ggpolt2
library(stringr) # Netejar dades
library(lubridate) # Format correcte dates
library(DT) # Visualització base de dades
library(robCompositions) # Mètodes d'imputació
library(compositions) # Detecció outliers, transformacions CLR, ILR i descriptiva
library(reshape2) # Gràfics
library(kableExtra) # Generació taules
library(knitr) # Generació taules
library(giscoR) # Generació mapes 
library(sf) # Generació mapes
library(terra) # Obtenció raster altitud
library(geodata) # Manipulació raster altitud
library(factoextra) # PCA
library(INLA) # Modelització amb INLA
